% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_postcode_vars.R
\name{add_postcode_vars}
\alias{add_postcode_vars}
\title{Add postcode variables e.g. Ward, to a data frame with postcodes}
\usage{
add_postcode_vars(
  df,
  pcd_name = "postcode",
  .admin_district = TRUE,
  .lat_long = FALSE,
  other_vars = character(0)
)
}
\arguments{
\item{df}{data frame with a postcode variable}

\item{pcd_name}{name of the postcode variable in passed the data frame}

\item{.admin_district}{boolean indicating whether to add an admin_district variable}

\item{.lat_long}{boolean indicating whether to add a pair of latitude & longitude variables}

\item{other_vars}{character vector of other variables to add}
}
\value{
data frame with variables added based on postcode
}
\description{
Add postcode variables e.g. Ward, to a data frame with postcodes
}
\examples{
# Create a data frame with some example records
df <- tibble::tribble(
  ~name,    ~postcode,
  "SCC",    "S1 2HH",
  "Blades", "S2 4SU",
  "Owls",   "S6 1SW"
)

add_postcode_vars(
  df,
  pcd_name = "postcode",
  .admin_district = FALSE,
  .lat_long = TRUE,
  other_vars = c("admin_ward", "msoa_code")
)
}
