% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrsheff-package.R
\docType{package}
\name{addrsheff-package}
\alias{addrsheff}
\alias{addrsheff-package}
\title{addrsheff: Tools To Help With Address Data}
\description{
For example, based on a postcode column, add variables to a data frame such as Ward.
}
\author{
\strong{Maintainer}: Laurie Platt \email{laurie.platt@sheffield.gov.uk}

}
\keyword{internal}
